+++ about.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Muse-ic - About</title>
+  <link rel="stylesheet" href="style.css">
+</head>
+<body>
+  <header>
+    <img src="img/logo-muse-ic.jpg" alt="Muse-ic logo" id="logo">
+    <nav>
+      <ul>
+        <li><a href="all-songs.html">All Songs</a></li>
+        <li><a href="artists.html">Artists</a></li>
+        <li><a href="about.html">About</a></li>
+      </ul>
+    </nav>
+    <div id="hamburger" class="hamburger" onclick="toggleMenu()">&#9776;</div>
+  </header>
+  
+  <main>
+    <h1>About</h1>
+    <p style="font-size: 24px;">Muse-ic is a music library application that allows you to manage your music collection.</p>
+    <p style="font-size: 24px;">You can add songs, search songs, sort songs and delete songs.</p>
+  </main>
+
+  <script type="module" src="about.js"></script>
+</body>
+</html>
