+++ about.js
@@ -1,0 +1,10 @@
+document.addEventListener("DOMContentLoaded", () => {
+  // Hamburger menu
+  function toggleMenu() {
+    const navList = document.querySelector("nav ul");
+    navList.classList.toggle("show");
+  }
+
+  // Hamburger event listener
+  document.getElementById("hamburger").addEventListener("click", toggleMenu);
+});
