+++ artists.html
@@ -13,7 +13,7 @@
       <ul>
         <li><a href="all-songs.html">All Songs</a></li>
         <li><a href="artists.html">Artists</a></li>
-        <li><a href="playlists.html">Playlists</a></li>
+        <li><a href="about.html">About</a></li>
       </ul>
     </nav>
     <div id="hamburger" class="hamburger" onclick="toggleMenu()">&#9776;</div>
