--- playlist-songs.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Muse-ic - Playlist songs</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <header>
-    <img src="img/logo-muse-ic.jpg" alt="Muse-ic logo" id="logo">
-    <nav>
-      <ul>
-        <li><a href="all-songs.html">All Songs</a></li>
-        <li><a href="artists.html">Artists</a></li>
-        <li><a href="playlists.html">Playlists</a></li>
-      </ul>
-    </nav>
-    <div id="hamburger" class="hamburger" onclick="toggleMenu()">&#9776;</div>
-  </header>
-  
-  <main>
-    <h1 id="playlist-name"></h1>
-
-    <div>
-      <img id="playlist-img" src="" alt="Playlist image" width="200" height="200">
-    </div>
-
-    <!-- Add song to playlist -->
-    <div id="add-song-container">
-      <input type="text" id="song-search" placeholder="Search song to add...">
-      <select id="song-select" size="5"></select>
-      <button id="add-song-btn">Add Song</button>
-      <p id="add-song-message" style="color:green; margin-top: 5px;"></p>
-    </div>
-
-    <!-- Search -->
-    <label for="search">Search:</label>
-    <input type="text" id="search" placeholder="Search songs...">
-
-    <!-- Sorting -->
-    <label for="sort">Sort by:</label>
-    <select id="sort">
-      <option value="last-added">Last added</option>
-      <option value="name-asc">Name A-Z</option>
-      <option value="name-desc">Name Z-A</option>
-      <option value="artist-asc">Artist A-Z</option>
-      <option value="artist-desc">Artist Z-A</option>
-      <option value="other-artist-asc">Other artist(s) A-Z</option>
-      <option value="other-artist-desc">Other artist(s) Z-A</option>
-      <option value="year-asc">Year old-new</option>
-      <option value="year-desc">Year new-old</option>
-      <option value="duration-asc">Duration low-high</option>
-      <option value="duration-desc">Duration high-low</option>
-    </select>
-
-    <!-- Table with all songs of the playlist -->
-    <div id="table">
-      <table>
-        <thead>
-          <tr>
-            <th></th>
-            <th>Name</th>
-            <th>Artist</th>
-            <th>Other Artist(s)</th>
-            <th>Year</th>
-            <th>Duration</th>
-          </tr>
-        </thead>
-        <tbody id="playlist-songs-list"></tbody>
-      </table>
-    </div>
-  </main>
-
-  <script type="module" src="playlist-songs.js"></script>
-</body>
-</html>
