--- playlists.html
@@ -1,48 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Muse-ic - Playlists</title>
-  <link rel="stylesheet" href="style.css">
-</head>
-<body>
-  <header>
-    <img src="img/logo-muse-ic.jpg" alt="Muse-ic logo" id="logo">
-    <nav>
-      <ul>
-        <li><a href="all-songs.html">All Songs</a></li>
-        <li><a href="artists.html">Artists</a></li>
-        <li><a href="playlists.html">Playlists</a></li>
-      </ul>
-    </nav>
-    <div id="hamburger" class="hamburger" onclick="toggleMenu()">&#9776;</div>
-  </header>
-  
-  <main>
-    <h1>Playlists</h1>
-
-    <!-- Add playlist form -->
-    <form id="playlist-form">
-      <input type="text" id="playlist-name" placeholder="Playlist Name" required>
-      <button type="submit">Add Playlist</button>
-    </form>
-
-    <!-- Search -->
-    <label for="search">Search:</label>
-    <input type="text" id="search" placeholder="Search playlists...">
-
-    <!-- Sorting -->
-    <label for="sort">Sort by:</label>
-    <select id="sort">
-      <option value="last-added">Last added</option>
-      <option value="name-asc">Name A-Z</option>
-      <option value="name-desc">Name Z-A</option>
-    </select>
-      
-    <div id="playlists-container" class="cards-container"></div>
-  </main>
-
-  <script type="module" src="playlists.js"></script>
-</body>
-</html>
